requirements.txt
streamlit
openai-whisper
pymongo
audio-recorder-streamlit
